+++ backend/src/main/java/hanglog/member/domain/auth/kakao/KakaoProvider.java
@@ -6,7 +6,7 @@ import static hanglog.global.exception.ExceptionCode.NOT_SUPPORTED_OAUTH_SERVICE
 import hanglog.global.exception.AuthException;
 import hanglog.member.domain.auth.Provider;
 import hanglog.member.domain.auth.UserInfo;
-import hanglog.member.dto.AccessTokenResponse;
+import hanglog.member.domain.auth.AccessTokenResponse;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
