--- backend/src/main/java/hanglog/member/provider/ProviderProperties.java
@@ -1,15 +1,0 @@
-package hanglog.member.provider;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public abstract class ProviderProperties {
-
-    protected final String clientId;
-    protected final String clientSecret;
-    protected final String redirectUri;
-    protected final String tokenUri;
-    protected final String userUri;
-}
